Support Characteristics keyword when defining Logical models (#1293)
* Add Characteristics keyword

Logical models can define Characteristics that indicate how they should
be used. Since there can be multiple characteristics, a list of codes
can be provided. Only allowed codes will be set on the logical model.
Any disallowed codes result in error messages.

* Set characteristics when exporting logical models

Logical model definitions with characteristics will include those
characteristics in the exported StructureDefinition's extension element.
The characteristics are checked against the type-characteristics-code
CodeSystem, and a warning is emitted if the characteristic is not
present in the CodeSystem. An exception is made for the "can-be-target"
characteristic, which was unintentionally left out of the CodeSystem. If
this CodeSystem is not available, a warning is emitted to let the author
know that their characteristics are not being checked.

* Fish for type characteristics codes once

* Only validate characteristics during export

Characteristic validation is removed from the importer. If an invalid
characteristic code is specified, a warning is logged during export. The
export warning lists the supported characteristic codes.

Remove hard-coded list of type characteristic codes. Use code system
definition to get list of type characteristic codes. Manually add
"can-be-target" to the list of supported codes if it is not present in
the code system. The list of supported codes is created in the
StructureDefinitionExporter constructor.

* Check referenced logicals for target validity

A logical model should specify the can-be-target characteristic for it
to be used as the target of a Reference. When constraining Reference
types, check to make sure that logical models that can be referenced
specify this characteristic. If there are any that do not specify this
characteristic, apply the type constraint and log a warning. When
assigning a specific Reference, check its type. If its type is a logical
model, and it does not specify the can-be-target characteristic, perform
the assignment and log a warning.

* Reduce amount of fishing when replacing references

When a referenced item is an Instance, we can usually get the metadata
for its type pretty early. Keep a reference to it for use in the logical
target check.

Try to set canBeTarget when fishing for metadata in FHIRDefinitions and
Package objects. If fishing for metadata for a logical model in a
Package, the extensions for indicating a valid reference target may not
exist if the export of that logical model is still in progress. In that
case, return no metadata to force fishing in the FSHTank as a fallback.

Move type characteristics and logical target URLs to fhirdefs to
simplify import structure.