Support looser type checking when constraining profile types
When constraining an element's type, the previous behavior required the new type to be a formal child of the old type. If the old type is a profile, however, the new type shouldn't have to be a formal subtype; it need only conform to the old type. When this occurs, log a warning but let it proceed.