Support caret value rule on value set concept
Elements within value set concept components can be set by caret value
rules. The concept must already be added to the value set in order to
use it in a caret rule. The concept can be in either the include or
exclude arrays. Concept rules with a single code establish a path
context that can be used by an indented rule that follows it.