System version doesn't need to match when assigning a code
A code assigned with a system may include a version as part of the
system. This is generally intended to refer to the system's published
version in the world, and is not guaranteed to match the version of the
CodeSystem FHIR resource. Fish for any version when resolving the
system, and do not emit a warning even if the versions don't match.