Allow caret and insert rules to be applied to concepts within a ValueSet (#1315)
* Support caret value rule on value set concept

Elements within value set concept components can be set by caret value
rules. The concept must already be added to the value set in order to
use it in a caret rule. The concept can be in either the include or
exclude arrays. Concept rules with a single code establish a path
context that can be used by an indented rule that follows it.

* Support insert rule on value set concept

Because caret value rules can be applied to concepts, it is also
reasonable to want to apply a rule set containing caret value rules at a
concept. This uses the existing grammar rule for a codeCaretRule, but
only one concept is allowed for the insert rule's path.

* Do not list more than one concept at a time for a ValueSet

The syntax that was used to support listing multiple concepts in a
single rule is removed. Tests related to this syntax are removed.

Add more tests for making sure that rule context is handled properly in
various combinations of rules on a ValueSet.

* Revert change to visitCaretPath

* Caret rule on ValueSet should have at most one concept