Set implied properties for all CaretValueRules at once in CodeSystemExporter and ValueSetExporter
Instead of setting implied properties one rule at a time each time
that setPropertyOnDefinitionInstance is called, collect all known
slices for all rules once and set all implied properties once. This
helps speed up performance because setImpliedPropertiesOnInstance is
expensive and causes code systems with a lot of CaretValueRules.

Right now, this change is only made in CodeSystemExporter and
ValueSetExporter but a similar approach can likely be taken for
StructureDefinitionExporter and other places
setPropertyOnDefinitionInstance is called.