Use title and name values in IG resource list (#1339)
* Use title and name values in IG resource list

Some resource types have title and name elements that are strings or
similar datatypes. If present, these can be used to set the name in the
IG resource list.

For package resources, only the title element value is used for this
purpose. A title set with a keyword will have higher precedence than a
title set by assigning to an element.

For predefined resources, the name element can be used for this
purpose as well. The title element has a higher precedence than the
name element. Note that the name element will only be used for
predefined resources, not package resources. This change applies to both
example and non-example predefined resources.

* Create utility function for string check