Rework approach to supporting R5 resources in R4/R4B
Rather than loading all of R5 as supplemental definitions and introducing special fishing modes, add only the necessary R5 resources to FHIRDefinitions when it is initialized -- so they are always fishable. Also add a _timeTraveler flag so we can appropriately reject these special resources where they're not allowed (namely as parents of profiles/logicals). Versions of FHIR that have these resources already simply overwrite the ones we added (and thus, do not have the _timeTraveler flag).

This approach simplifies the code a good bit since we can continue to fish the normal way and we don't need to work out as many edge cases. It's also more efficient since we're not loading all of R5 in memory. Last, it's more consistent w/ the approach we already used for supporting the R5 Base Resource.