Support Instances of Allowed R5 Resources in R4/R4B IGs (#1350)
Allow authors to create instances of R5 ActorDefinition, Requirements, SubscriptionTopic, and TestPlan in R4 IGs - and refer to them via references and canonicals (e.g., from an obligations extension).

This is primarily supported by adding these special resources to FHIRDefinitions when it is initialized -- so they are always fishable. We also add a _timeTraveler flag so we can appropriately reject them where they're not allowed (namely as parents of profiles/logicals). Versions of FHIR that have these resources already simply overwrite the ones we added (and thus, do not have the _timeTraveler flag).

Other changes to support this: MasterFisher now determines its default version of FHIR when it is initialized. I've also added a FHIRVersionsUtility and refactored some existing code to leverage it.