Maintain children of meta.profile when removing duplicates
Values of meta.profile are not allowed to be duplicated. When duplicates
are detected, check if the removed entries have any child elements. If
they are present, merge them into the entry that it matched. This mostly
applies to extensions, which have their elements concatenated. For
primitives, such as id, the value on the non-duplicate element is
preferred.