Don't return instance rules when looking for id (#1355)
* Don't return instance rules when looking for id

When searching for a rule that sets an id, check the rule's isInstance
flag. Only return the rule if the flag is not set.

* Handle instances when getting values from FSH rules

When getting the URL or version from FSH rules, only use the value if it
is actually just a string, and not an Instance.

Remove getRuleValue function, which was just returning the value
property.