Improve handling of child elements on meta.profile (#1353)
* Improve handling of child elements on meta.profile

An Instance may have extensions or other child elements defined on
meta.profile. When these elements exist, they should not interfere with
checking for the InstanceOf within the profile array. Additionally, if
values are set on meta.profile and on child elements, keep those values
at aligned indices when adding InstanceOf to meta.profile.

* Maintain children of meta.profile when removing duplicates

Values of meta.profile are not allowed to be duplicated. When duplicates
are detected, check if the removed entries have any child elements. If
they are present, merge them into the entry that it matched. This mostly
applies to extensions, which have their elements concatenated. For
primitives, such as id, the value on the non-duplicate element is
preferred.