Consider imposeProfile extensions when constraining types (#1366)
* Consider imposeProfile extensions when constraining types

Profiles can declare that they also require conformance against other profiles via the imposeProfiles extension. For example, Profile A might declare imposeProfiles X and Y. This means that wherever X or Y are allowed, A should be allowed too (since it requires conformance to X and Y).

When constraining an element type, SUSHI will now consider declared imposeProfiles in addition to the type's type hierarchy. This is made simpler by adding imposeProfiles to the metadata that can be fished up.

This commit also improves detection of special extensions in the tank when those extensions use the named extension syntax instead of numeric indices.

* Refactor findImposeProfiles to a common module

* Remove unnecessary commented out line