Try to avoid overwriting child values of implied CodeableConcept
When a Profile has an assigned pattern value on a CodeableConcept
element, the coding child element may have one or more values. If an
Instance of that Profile assigns to that coding element, try to avoid
overwriting the coding from the pattern with a value that does not match
the pattern. An assignment that matches the pattern should also be fine.