Create index files for FSH-defined entities
There is one entry in the index for each FSH definition that results in
an exported JSON file. Inline instances do not appear in their own file,
so they do not appear in the index. The index is created in two formats.
There is a txt file that is formatted for human readability, and there
is a JSON file that is useful for machine processing.