Produce correct FSH for concept CaretValueRules on ValueSets (#1385)
* Produce correct FSH for concept CaretValueRules on ValueSets

When a CaretValueRule on a ValueSet has a concept as its path, the
system is also included. However, this is not the case when the rule is
on a CodeSystem. CaretValueRule should produce FSH that is correct in
both cases.

* Include # in path array when visiting Concept

Preserving the # allows the path array to be used with indented caret
rules.

When producing FSH for a ConceptRule, the code part needs to be quoted
if it contains spaces or starts with a double quote character.

* Quote codes in CaretValueRule paths that start with quotes

---------

Co-authored-by: Chris Moesel <cmoesel@users.noreply.github.com>