Allow binding to inline and example ValueSet instances
A binding rule can use an inline ValueSet as long as the ValueSet is
contained within the resource where the binding is being applied. The
binding will use a relative reference. Binding rules that refer to
inline instances are deferred until after caret rules are applied.

A binding rule can use an example ValueSet. This works the same as using
a definitional ValueSet.