Track binding rules and replace URLs of contained ValueSets
When binding an element on a StructureDefinition to a ValueSet, always
prefer using a reference to a contained ValueSet rather than the
ValueSet's url. This is true regardless of how the ValueSet is defined.

Emit an error when binding to a ValueSet defined as an inline instance
that is not contained within the StructureDefinition where the binding
is applied.