Fish in tank by name when set by a rule
ValueSet and CodeSystem are both defined with a name element. If a rule
is used to set the value of this element, that value should allow you to
fish for the ValueSet or CodeSystem. This also applies to Instances of
ValueSet and CodeSystem.

Refactor some functions that find rules at paths to reduce the amount of
nearly-duplicate functions. Some of the path-specific functions are kept
if they have additional logic.