Find instances in the FSH tank by assigned name
Some resource types have a "name" element. If an Instance in the FSHTank
has a non-instance value assigned at this path, the Instance can be
fished up by searching for that value.

This mostly applies to CodeSystem and ValueSet resources. When fishing
in the tank for a CodeSystem or ValueSet, these Instances can already be
found. But, there are some cases where SUSHI is specifically fishing for
an Instance, and it still needs to find these.