Do not fish up Instances in FSHTank by value of name
Undo previous change that allowed fishing by a value assigned to the
name element on an Instance. This was something of a workaround to
MasterFisher's implementation of trying to find the resource type for an
Instance in the tank. Instead, MasterFisher now fishes for the
originally requested types, and if an Instance is found, it uses the
Instance to set the resourceType.