Fish for specific types when finding SD parent and type lineage
The parent of a StructureDefinition can only be a Profile, Extension,
Resource, Logical, or Type. Similarly, when building a type lineage,
each type will be a Profile, Extension, Resource, Logical, or Type. Type
is left out when fishing in the tank, since they can't exist in the
tank. This removes the need for ParentNameConflictError.