Allow binding to inline ValueSet instances (#1395)
* Allow binding to inline and example ValueSet instances

A binding rule can use an inline ValueSet as long as the ValueSet is
contained within the resource where the binding is being applied. The
binding will use a relative reference. Binding rules that refer to
inline instances are deferred until after caret rules are applied.

A binding rule can use an example ValueSet. This works the same as using
a definitional ValueSet.

* Track binding rules and replace URLs of contained ValueSets

When binding an element on a StructureDefinition to a ValueSet, always
prefer using a reference to a contained ValueSet rather than the
ValueSet's url. This is true regardless of how the ValueSet is defined.

Emit an error when binding to a ValueSet defined as an inline instance
that is not contained within the StructureDefinition where the binding
is applied.

* Fish in tank by name when set by a rule

ValueSet and CodeSystem are both defined with a name element. If a rule
is used to set the value of this element, that value should allow you to
fish for the ValueSet or CodeSystem. This also applies to Instances of
ValueSet and CodeSystem.

Refactor some functions that find rules at paths to reduce the amount of
nearly-duplicate functions. Some of the path-specific functions are kept
if they have additional logic.

* Find instances in the FSH tank by assigned name

Some resource types have a "name" element. If an Instance in the FSHTank
has a non-instance value assigned at this path, the Instance can be
fished up by searching for that value.

This mostly applies to CodeSystem and ValueSet resources. When fishing
in the tank for a CodeSystem or ValueSet, these Instances can already be
found. But, there are some cases where SUSHI is specifically fishing for
an Instance, and it still needs to find these.

* Do not fish up Instances in FSHTank by value of name

Undo previous change that allowed fishing by a value assigned to the
name element on an Instance. This was something of a workaround to
MasterFisher's implementation of trying to find the resource type for an
Instance in the tank. Instead, MasterFisher now fishes for the
originally requested types, and if an Instance is found, it uses the
Instance to set the resourceType.

* Fish for specific types when finding SD parent and type lineage

The parent of a StructureDefinition can only be a Profile, Extension,
Resource, Logical, or Type. Similarly, when building a type lineage,
each type will be a Profile, Extension, Resource, Logical, or Type. Type
is left out when fishing in the tank, since they can't exist in the
tank. This removes the need for ParentNameConflictError.

* List types for fishing in default order