Assign FSH Code or Reference to CodeableReference element
When writing an assignment rule, if the path refers to a
CodeableReference element, the value may be a FshCode or FshReference.
The value will be applied to the element's concept or reference child
element, respectively. It is possible to have two rules with the same
CodeableReference path if one assigns a FshCode and the other assigns a
FshReference.