Cache StructureDefinitions used by InstanceExporter
When exporting an Instance, the instanceOf StructureDefinition is used
to validate rules. Rule validation may require unfolding elements, which
can be an expensive operation. By caching the StructureDefinition, those
operations will only need to be performed once for the
StructureDefinition, regardless of how many Instances of that
StructureDefinition are exported.