Track parent/child relationships of ElementDefinitions
An ElementDefinition maintains a reference to the element that is its
parent. It also maintains a list of references to its immediate
children. When adding an element, set its parent, and add it to its
parent's list of children. Functions to return an element's parent or
children will use these references when they are available.