Combine included concepts when exporting ValueSet (#1432)
* Combine included concepts when exporting ValueSet

ValueSet concepts with the same system and version may appear as
separate rules. This happens most often when RuleSets are used. Included
concepts with the same system and version are added to the same concept
list.

* Check valuesets before combining concepts

For concepts to be combined, the system, version, and valuesets must all
match. The order of valuesets does not need to be the same.

Combine concepts on the exclusion list.