Improve performance when exporting Instances (#1443)
* Cache StructureDefinitions used by InstanceExporter

When exporting an Instance, the instanceOf StructureDefinition is used
to validate rules. Rule validation may require unfolding elements, which
can be an expensive operation. By caching the StructureDefinition, those
operations will only need to be performed once for the
StructureDefinition, regardless of how many Instances of that
StructureDefinition are exported.

* Track parent/child relationships of ElementDefinitions

An ElementDefinition maintains a reference to the element that is its
parent. It also maintains a list of references to its immediate
children. When adding an element, set its parent, and add it to its
parent's list of children. Functions to return an element's parent or
children will use these references when they are available.

* Rebuild element tree after cloning Logical/Resource elements

When defining a FSH Logical or Resource, the initial set of elements is
created by cloning the parent's elements. After cloning these elements,
rebuild their tree relationships.

* update comments for clarity and accuracy

* Improve performance of tests that compare ElementDefinitions

* simplify slicing arguments