Assign date or dateTime that contains only year
A FHIR date or dateTime may be 4-digit year. The FSH parser will assign
this value a NUMBER token type. When assigning to a date or dateTime
element, if a type mismatch occurs, and the original value is a number,
try to assign the raw value. If it is a valid 4-digit year, it will be
assigned successfully. Otherwise, the original error will be thrown and
logged.