Only return FshType id from rule value if the value is a string (#1524)
When getting the id for a FshType, rules are checked to see if there is
one that will set the id. A value that is not a string will not be valid
and not be applied. Therefore, only return the value if it is a string.

Rename test file for FshCodeSystem to match name of source file.