Allow ValueSet to reference contained inline CodeSystem (#1520)
* Allow ValueSet to reference contained inline CodeSystem

If a ValueSetComponentRule specifies a system, check if that system is
present in the list of contained resources. If so, add the
valueset-system extension to the system, using the relative reference as
the extension's value. If the system is not present in the list of
contained resources, and the system is an inline Instance, log an error
and do not add the component to the ValueSet.

Fishing in the tank for a CodeSystem will now return inline instances of
CodeSystem. This matches the operation of fishing in the tank for a
ValueSet.

Add _system.extension to the type definition for elements of a
ValueSet's include and exclude lists. Use the Extension fhirtype as part
of this definition. Replace the Extension fshtype with the Extension
fhirtype in other parts of the ValueSet definition.

* Better type for ValueSet._system

Add link to zulip thread with information about the contained code
system extension.

* Improve handling case when referenced system is not found.

Add a test showing that an inline example instance of a CodeSystem will
not work.
Add a currently-skipped test that shows what the inline ValueSet
behavior should be, based on currently available information from this
zulip thread:
https://chat.fhir.org/#narrow/stream/215610-shorthand/topic/Contained.20code.20system.20in.20the.20value.20set/near/475597931