Support fully defining resources using caret rules
A contained resource can be defined within another resource one rule at
a time using caret rules. By setting the resourceType of the contained
resource, SUSHI will be able to correctly validate other rules that
build up the contained resource. Most of the logic for this already
existed, but some additional code was needed in order to collect
information about inline resource types. Additionally, if an instance is
assigned with a caret rule, further caret rules may be used to modify
the contained resource. This gets a little complicated when exporting a
StructureDefinition.

A ValueSet may refer to a CodeSystem that is defined with caret rules on
the ValueSet. The resulting compose element will look the same as if the
inline CodeSystem were defined separately.

Log a warning when an example instance is assigned within a conformance
resource.

Update object definitions for inline instance tests so that the
Instances are, in fact, inline.