Support "position" slicing discriminator
FHIR R5 added "position" as an allowed slice discriminator, where slices
are differentiated by their order within a list element. Because a
consistent order is necessary for this slicing to be meaningful, emit an
error if "ordered" is not set to true for the slicing. Previous versions
of FHIR may not use this discriminator.

Minor changes to npm-shrinkwrap.json due to issues detected by `npm
audit`.