Integrate FPL 2.0.0-beta.2
The main difference in FPL 2.0.0-beta.2 is that you no longer pass in a sql.js Database to the SQLJSPackageDB constructor. Instead, you construct SQLJSPackageDB and then call its async initialize function.

This required us to update FHIRDefinitions to add an initialize function that then invokes the SQLJSPackageDB initiailze function.