Check exported FHIR for multiple choice values
Only one type choice of a choice element should be present on exported
FHIR. Check for the presence of multiple type choices. If multiples are
found, log an error. Note that it is still fine to apply rules to
different type choice elements on a Profile, since those are two
separate element definitions in the snapshot and differential lists.