Check exported resources for presence of multiple choice values (#1537)
* Check exported resources for presence of multiple choice values

When a choice element is defined in FHIR, it allows for values of
different types with names based on those types. However, any choice
element that is present on resource should have only one named value.
For example, Observation.value[x] allows for many types, but an
Observation resource should not have both valueString and valueInteger.
After exporting a resource, check the resource for the presence of
multiple values for the same choice element, and log an error when they
are detected.

As assigned values are validated in exporters, the relevant
StructureDefinitions will have their elements unfolded. Keep the same
instance of that StructureDefinition that was used for validation so
that the unfolded elements are present when performing the multiple
choice validation check.

* Check exported FHIR for multiple choice values

Only one type choice of a choice element should be present on exported
FHIR. Check for the presence of multiple type choices. If multiples are
found, log an error. Note that it is still fine to apply rules to
different type choice elements on a Profile, since those are two
separate element definitions in the snapshot and differential lists.