Allow fragment references in Canonical, Reference, and system
A fragment reference may be used with the Canonical or Reference
keywords when referring to a contained resource. A fragment reference to
a CodeSystem may be used when defining a ValueSet.

The current implementation is not yet complete. There is still a need
to support fragment references to contained ValueSets when defining a
ValueSet. Additional tests may also be needed for the existing
functionality.