Allow base types as profiles in type constraint rules (#1589)
Provides the ability to specify both a profile and the base type as a given only-rule. Previously, when calculating the intersection between types and profiles, if the profile === type, it was just dropped. Now it will be included as an additional profile, but only if it was not the only instance of that type.